Add files via upload
This is a detailed dashboard of the IPL analysis. This dashboard is made with the large dataset of all the season of IPL. From 2008 to 2022. 
The dashboard is made by cleansing the data in the necessary columns . The use of DAX function is also done to get the more accurate insights in the report.
Various KPIs are highlighted in the top to get the necessary information in the get go. The dashboard is completely interactive with slicers which can be used to filter the different seasons . And 2 separate slicers are also given to highlight the detailed stats of batsman and bowlers individually . User can also search the individual stats of  batsman and bowler in the slicer to get the detailed stats of the desired cricketer.